1.7.2 - I Like Trains
# Changes

- Slightly buffed Furnace Minecarts. (50% speed increase)

- Fixed Debug Sheep Drops

# Why?

This may seem like such a radical change, but it is really miniscule; furnace minecarts are obsolete as-is--even in this version where there are no powered rails. I wanted to give the furnace minecarts the true train engine power they deserve.. I had a vision of the polar express in my winter world while making this..

I believe this comes to no compromise of the vanilla experience.. furnace minecarts should have always been this speed.

Good Day.