1.7.2_02 - Rail Crash Fix
Addresses Issue OAP-3.

# Bug Fix

When placing a block, it had ran the method "onBlockPlacedBy" for every block that was placed. For rails (and probably other things that still rely on neighbor updates to rotate), they could not sufficiently use that method, where it threw the NoSuchMethodError. I changed the logic where the "onBlockPlacedBy" method is triggered if and only if the block is a furnace or a stair (Blocks that needed player-oriented rotating), resolving the crash.

Good Day.