1.7.4 - Zooming
# Additions

Allowed for the ability to zoom with G.

Allowed the ability to change the key bindings for zooming and perspective, with their default binds being G and F5, respectively.

# Changes

The user no longer has to hold down F3 to open the debug menu, they can just press the key.

Good Day.